Silently skip the Wincache tests on travis to reduce build noise
It's unlikely travis will ever be able to test this class, because
it only runs on windows. The big block of skipped tests creates a
lot of noise and will mask skipped tests that we actually care
about.

Therefore, on travis we skip the test class definition to force
silent skipping of the tests.