Test ignore_on_delete config setting on an arbitrary file
When unit-testing a Kohana application the conventional way, without
Koharness and with the cache tests enabled, the .gitignore file in the
`application/cache` folder ends up deleted.

With this change, tests does not affect `.gitignore` but run on an
arbitrary `file_we_want_to_keep.cache` file that gets created while
testing, and then destroyed.